// api/src/types/apiContracts.ts
export interface WsCacheInvalidatePayload {
    path?: string;
    key?: string;
    target?: string;
    timestamp: number;
}

export interface WsVoiceUpdatePayload {
    userId: string;
    guildId: string;
    channelId: string | null;
    timestamp: number;
}